Moved commit message building to `CommitMessage`.
What
===
Moved all commit message building logic into `CommitMessage`,
changing the existing `#joined_message` so that it does not include the
author and created `#joined_message_with_author` that does. `GitSvn` and
`Svn` now use the latter and `Git` uses the former.

Why
===
So that the joining and formatting of commit messages is entirely done
in one location. Its tells a clearer story about how the commit message
is converted into a single string.