Removed `puts` from `PromptLine#message`.
What
===
Removed the `puts` lines from `PromptLine#message` and removed the
method all together.

Why
===
It was a little odd that the `message` function had dual purpose, to
display the previous input as well as return the question to output for
the user.

The method was removed once it only contained simple interpolation and
did not make the code clearer by keeping it.