Refactored the `PromptLine` history load and set.
What
===
Refactor the `PromptLine` history loading and setting by:

* Moving the logic that setup the `Readline` history into a method
named `set_readline_history`. This logic previously existed in
the `load_history` method.
* Moving the logic that loaded the history out of file into a method
named `load_history`. This logic previously existed in the
`saved_history` method.

Why
===
To improve clarity of the intention of functions and what they do. To
separate the reading in of the history (now in `load_history`) with that
history being used internally, and the use of that loaded history to
setup the `Readline` history (now in `set_readline_history`).
Previously the return value of the setup of the `Readline` history was
also used as the input to the internals, which was unnecessary.