Removed Gemfile.lock.
What
===
Removed Gemfile.lock.

Why
===
It should not be committed in gems, "since it enforces precision that
does not exist in the gem command, which is used to install gems in
practice. Even if the precision could be enforced, you wouldn't want it,
since it would prevent people from using your library with versions of
its dependencies that are different from the ones you used to develop
the gem." Ref http://yehudakatz.com/2010/12/16/clarifying-the-roles-of-the-gemspec-and-gemfile/.