Removed .ruby-gemset and .ruby-version
What
===
Removed .ruby-gemset and .ruby-version.

Why
===
These files are intended to be used for standalone projects I think and
not for gems since the gemset shouldn't define the gemset it's included
in. Additionally the ruby version should be whatever version the parent
project is using and shouldn't be defined by the gem.